# Optimize the HStore parser