### Context

I saw some people reporting poor Hstore performance on Reddit, and after taking a quick look, it indeed seems that the parser is quite wasteful.

### Compatibility

A good part of the extra performance is at the cost of making the parser less tolerant. Maybe extra tests based on the postgres documentation were added by @joelhoffman in https://github.com/rails/rails/pull/4896, but from my understanding, the documentation states what postgres will accept, not what it will produce.

e.g. unquoted `NULL` is case insensitive, so postgress will accept `NuLL` instead, but will always return `NULL`.

I'm not experienced with postgresql though, so I'm might be totally off.

### Performance

Test string from taken from the the test suite: `"\"a\\\\b\"=>\"b\\\\ar\", \"1\\\"foo\"=>\"2\""`

Bench:

```
Warming up --------------------------------------
            original     6.896k i/100ms
             patched    15.787k i/100ms
Calculating -------------------------------------
            original     68.176k (± 1.2%) i/s -    344.800k in   5.058270s
             patched    157.786k (± 1.0%) i/s -    789.350k in   5.003144s

Comparison:
             patched:   157786.0 i/s
            original:    68176.1 i/s - 2.31x  (± 0.00) slower

```

Memory before: allocated 4376 bytes (55 objects)
Memory after: allocated: 1880 bytes (23 objects)