@casperisfine 2021-04-13
Thanks to @pixeltrix I pushed another version using `StringScanner` that's even slightly faster: https://gist.github.com/pixeltrix/a9aafda5d4d8bedf709ff54636551986

I'll merge on green.