@pixeltrix 2021-04-13
@casperisfine did you leave the `HstorePair` constant in for a reason?